requirements.txt
streamlit>=1.25.0
fastapi>=0.68.0
uvicorn>=0.15.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pydantic>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
gunicorn>=20.1.0
